pyproject.toml
[project]
name = "the-bois"
version = "0.1.0"
description = "Local multi-agent LLM system — the bois got your back."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "ddgs>=7.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pytest>=8.0.0",
]

[project.scripts]
the-bois = "the_bois.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
